Sound/Music/Gameplay stuff
Fixed merge conflict on my end.
Added code so that the game genres for all objects are tied to the gameManager's game genre.
Moved Audio folder to Resources folder for easier access.
Sounds now play when jumping and/or shooting.
Shooting can only be done when genre in GameManager is shooting.
Music now plays based on the current genre.